LPS-56211 move
