+++ portal-impl/src/com/liferay/portlet/blogs/image/selector/BaseBlogsImageSelectorUploadHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.blogs.action;
+package com.liferay.portlet.blogs.image.selector;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
